feat: improve app persistence, ui cleanup, mini mode, and api modernization
- app minimizes to tray instead of closing
- removed redundant options from general tab
- implemented control panel mini mode (collapsible)
- modernized win32 apis: upgraded to Reg...Ex functions and GetTickCount64
- updated audio engine and ui components for 64-bit timing compatibility

+++ src/audio/StreamingWavWriter.h
@@ -61,7 +61,7 @@ private:
     std::atomic<bool> m_failed;
     std::atomic<size_t> m_totalBytesWritten;
 
-    std::atomic<DWORD> m_lastFlushTime;
+    std::atomic<ULONGLONG> m_lastFlushTime;
     void PeriodicFlush();
 
     // Audio format
